Feature: Update and delete card

+++ backend/AlfaWallet/src/main/java/com/alfawallet/backend/model/repository/CardRepository.java
@@ -1,11 +1,10 @@
 package com.alfawallet.backend.model.repository;
 
 import com.alfawallet.backend.model.data.CardData;
-import com.alfawallet.backend.model.data.UserData;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardRepository extends JpaRepository<CardData, Long> {
+public interface CardRepository extends JpaRepository<CardData, Integer> {
 
 }
